Removed other dangerous race condition, by not untransforming the plot_context in our display loop.
Without this there was a chance that points/lines would be drawn to the
wrong places on the plot surface. The main race condition that still
exist (transform_to_plot_units) could result in wrong line_widths etc,
since the other thread expects to be drawing to a untransformed context.
But shouldn't actually cause wrong positions, just ugly shapes.