Use a temporary surface to plot to and then plot that surface to the xlib surface (screen)
Solves the flickering problem, although it is probably a bit slower.
Being able to use axes as a mask would also have solved it, but we would
need to invert the alpha of that surface. This is a simple fix, with low
performance cost probably.

Also removed this task from the todo list