Closing plot window won't crash the calling application
Events are from then on just ignored

There might be an race condition when pBPlot becomes NULL during the
process_events loop (i.e. inside the while) but I have been unable to
trigger this. This is actually indeed probably not the case, because the
close window is an event and is therefore called from within the loop,
i.e. not from another thread. So we should be safe.