Fixed the zoom/rolling update bug
Bug was caused by the fact that the new surface (create_plot_surface)
would be relatively bigger (in user area) than the old surface, while
remaining the same pixel count. This was because create_plot_surface
used the (due to zooming) changed max_x etc values to create the new
surface. Decided to calculate the user area only at start time, which
would then not be recalculated again by create_plot_surface. The other
solution would lead to scaling issues.

This does mean that zooming will need to be limited to not be larger
than the original area set aside.