Committed fix for some updates not immediately showing up
This could happen when the last event in the event queue didn't call
BackendPlot::display() and therefore the screen was not refreshed.
Currently fixed by always calling display when the queue happens to be
empty. Not ideal (if the last event did call display we now call
display twice), but best I think I can do.