Enabled -Wall by default and cleaned up warnings
One set of warnings was caused by often using conf->window instead of
the correct event->event (which should point to the window of the
event). I guess this didn't cause problems because "activating" a window
would cause a configure event and therefore cause the next events to go
to the correct window anyway.

Many cleanups were just reordening of the initialization of variables to
reflect the actual order they would be initialized in